reorganize fab-related files into fab/ directory
