+++ fab/pillows/staging.yml
@@ -5,4 +5,3 @@
 # Add to this to prevent groups of pillows from running
 pillowtop_blacklist:
   - fluff
-
