+++ fab/services/templates/supervisor_celery_beat.conf
