+++ fab/services/templates/supervisor_celery_doc_deletion_queue.conf
