+++ fab/services/templates/supervisor_celery_flower.conf
