+++ fab/services/templates/supervisor_celery_main.conf
