+++ fab/services/templates/supervisor_celery_periodic.conf
