+++ fab/services/templates/supervisor_celery_pillow_retry_queue.conf
