+++ fab/services/templates/supervisor_celery_reminder_queue.conf
