+++ fab/services/templates/supervisor_celery_reminder_rule_queue.conf
