+++ fab/services/templates/supervisor_celery_sms_queue.conf
