+++ fab/services/templates/supervisor_couchdb_lucene.conf
