+++ fab/services/templates/supervisor_django.conf
