+++ fab/services/templates/supervisor_elasticsearch.conf
