+++ fab/services/templates/supervisor_formsplayer.conf
