+++ fab/services/templates/supervisor_pillowtop.conf
