+++ fab/services/templates/supervisor_sms_queue.conf
